MEN-2939: Make rootfs updates faster
This is done through reading one block ahead from the payload stream, and
comparing it with the data already present on the block-device. If the data
matches, simply skip ahead to the next block.

This should significantly increase write speed, as on most IO devices, reading
is significanlty faster than writing.

Changelog: Commit

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>