fix(testcoverage): Disable the paralell run of TestMutualTLSClient
This significantly raises the test coverage for the two affected files which had
their coverage lowered by ~8% by the previous commits. To me this signifies that
our test-framework does not handle the parallell test coverage properly. I tried
running it with -covermode atomic, but then the tests crash :/ It is a little
out of scope to look into this now.

See:

https://tracker.mender.io/browse/MEN-3881

for follow up.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>