Add: Make the dirstore able to read absolute file names
Previously it was always assumed that the `mender-agent.pem` file was present in
`/var/lib/mender` on the data partition, and hence the filepath was always
assumed to be absolute. This assumptions is voided by the new need to load keys
and certificates for the mTLS client functionality from the configuration file.

Therefore, make sure that the dirstore does not join paths when presented with
an absolute path to load a file from.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>